refactor(scripts/i18n): move region models based code into a separate file
