+++ scripts/i18n/cli.js
@@ -5,6 +5,7 @@
 const yargs = require('yargs')
 
 const utils = require('./utils')
+const regions = require('./regionCommons')
 
 const colors = {
 	reset: '\x1b[0m',
@@ -95,7 +96,9 @@ const getArgs = (description, opts) => {
 		args = args.option('file', {
 			alias: 'p',
 			type: 'string',
-			description: `The source file to translate from the 'locales/pages' directory. If not specified, all the files in 'locales/pages' will be translated.`,
+			description:
+				opts.file.description ??
+				`The source file to translate from the 'locales/pages' directory. If not specified, all the files in 'locales/pages' will be translated.`,
 		})
 	}
 	if (opts.remove) {
@@ -147,6 +150,17 @@ const getArgs = (description, opts) => {
 		return l !== srcLang
 	})
 
+	const destRegions =
+		argv.model?.filter((r) => {
+			if (!regions.supportedRegionCodes.includes(r)) {
+				cli.printWarn(
+					`[WARN] - the region '${r}' is not supported, skipping it.`
+				)
+				return false
+			}
+			return true
+		}) ?? regions.supportedRegionCodes
+
 	const srcFile = argv.file ?? opts.defaultSrcFile
 
 	return {
@@ -155,7 +169,7 @@ const getArgs = (description, opts) => {
 			!argv.target && opts.target === 'all'
 				? utils.availableLanguages
 				: destLangs,
-		destRegions: argv.model,
+		destRegions,
 		force: argv.force,
 		remove: argv.remove,
 		srcFile,
