+++ scripts/i18n/utils.js
@@ -9,6 +9,10 @@ const LOCK_KEY_EXT = '.lock'
 const AUTO_KEY_EXT = '.auto'
 const PREVIOUS_REVIEW_KEY_EXT = '.previous_review'
 
+const publicDir = path.resolve('public')
+
+const t9nDir = path.resolve('data/i18n/t9n')
+
 const availableLanguages = ['fr', 'en-us'] //, 'es', 'it'] For now, we don't want es and it to be compile (it could create compilation errors).
 const defaultLang = availableLanguages[0]
 
@@ -296,4 +300,5 @@ module.exports = {
 	objPath,
 	assoc,
 	customAssocPath,
+	publicDir,
 }
